Extend in what circumstances the env is updated
In order to directly support the ability to open a Terminus terminal in
either a tab or a view and have its environment automagically updated,
this commit introduces a setting that allows for specifying what
commands the environment should be set for.

In addition, the code for updating the environment was extended so that
it always starts from a pristine copy of the previously saved
environment, since if you use `terminus_exec` as a target in a build,
the command sequence is: build -> terminus_exec -> terminus_open, and
since we want to watch terminus_open, it will also get the environment
update.

Finally, this fixed a really dumb bug in the original version of the
plugin where we forgot to update the name of the command invoked when
we restore the environment, and the logging in the command that sets
and restores the environment is now much more verbose so it is directly
obvious what caused updates and where they are happening.