Include a fetched configuration cache
Implement and tie in an environment cache mechanism that acts as the
storage for keys requested from Envault for any given configuration
during the current run.

This is used when setting the environment, but does not yet get
populated by requests.

Note also that this current version simplistically uses the window ID of
the value, but could (and probably should) use the filename of the
configuration file loaded as the key.