Include a command to select and load a config file
Implement a new basic command which scans for the list of config files
that exist in the project and requests that the user select one. The
selected configuration file is loaded and the contents are used to
query Envault for the appropriate values.

This currently does not handle situations like there not being any
config files to choose from.