Update README and bump the mkdocs version
As first steps towards actual documentation for the package, replace the
stub project README with one that has a few extra details.

In addition, in spinning up the documentation templates originally I had
to bump the mkdocs-material version up since the one we were using ran
into issues with Google fonts and stopped working.

After that upgrade, a deprecation warning started being thrown by the
git revision plugin, which we left alone as harmless.

Since documentation work is actually spinning up now, it makes sense to
actually resolve that, which is done by simply bumping up the version on
that plugin so that it no longer takes the deprecated action.