Swap the environment cache to use filename
The idea here is that caching by window ID is bad because it's possible
for the same project to be open in multiple windows via multiple
workspaces, and in that case we want them all to share the same
environment.

So, instead of caching the environment per window id, we cache it by
the filename that produced it instead.

In addition this fixes an issue in which an empty environment cannot be
fetched for a window when there is no config present unless the debug
flag is turned on.