Gracefully handle selecting when no configs exist
Previously the code in the input handler was blowing up if there were
no configs found because it tried to select a default item when there
were none.

This change resolves that by having it detect that situation; this will
then cause the command to be immediately invoked with `None` as the
config selected, allowing the command to bail.

As a part of this, fix an issue with the log function, which was not
properly handling the case where the log should be a status message.