Clean up the display in the envault config chooser
Previously the command to choose a configuration file for the window
would display just the full absolute path names for the Envault config.

In order to make that easier to see, it now displays primarily just the
names of the config files, and as an extra layer shows you the folder
that contains that particular config, for more context.