On plugin load, fetch all configs currently in use
When the plugin loads (which includes at Sublime startup), scan over all
of the windows that were restored to see if they have set within them a
currently selected Envault configuration file.

For any that do, kick off a request to get the variables for that
config, so that upon restart of Sublime all of the required data to keep
tooling work is available.