Include a command to reload the active config
Although it is possible to reload the current config by just selecting
it again from the list, in cases where there is more than one config
present it is handy to be able to reload it directly, since there is no
direct way to know which one is currently active.

This implements such a command, which is only available in windows that
have a currently selected configuration.