Implement a command to show the current variables
Just because it might be handy, implement a simple command which shows
the list of environment variables that are currently being applied in
the window by Envault; this also shows you via the placeholder in the
quick panel what the loaded config file actually is.

This shows only the variables and not their values for security reasons.