Update the config cache to use explicit filename
The previous cache mechanism was based on passing in the window so that
we could use it's ID, which was then altered to use the config filename
set for that window.

In order to clarify further, the actual API now directly takes a config
filename instead, since it's not actually per window.

All call points are now set up to call in with the filename instead now
as a result.