Reuse existing environments from the cache
Implement actual cache-like behaviour for the environment cache.
Previously it was used more as a key-value storage directly, but it is
now possible to ask it if it contains environment for a specific config
or not.

This allows true re-use, such as not having to re-fetch the config if
you close and reopen a project in a specific session, or when a project
spreads over multiple windows.