Clarify the config loading code
As we will want to be able to not only find and load config files but
also create them, adjust the module name so that it makes more sense.

In addition, the extension to be used on the config files was hard coded
when it should not be; this has also been fixed.