Clarify Envault request failures
Since there may be potentially many requests in flight across windows at
the same time, make it easier to determine which one failed by
displaying the URL that had the error.

This could be further augmented to display the name of the config file
that was being loaded at the time if that is deemed interesting.