Include a command to create a new config
Scaffold out a command which can be used to create a new configuration
file for Envault based on some configured template values.

Care is taken not to clobber an existing configuration without first
asking the user, and it is possible to activate the config immediately
after it is created.

The newly created configuration file is opened so that the user can edit
it. A new event listener allows for the file to dynamically reload when
you change it.