Display current config in the status bar
This commit scaffolds out the code required to display in the status bar
an indication of what the currently selected configuration in any given
window is.

This is applied at startup to all views as configs load, every time a
file is loaded or cloned, when a new file is created, and also when a
file moves between windows.

In the cast of the latter, we do a little bit of jiggery-pokery here
because the on_post_move event does not give us the appropriate window,
which makes us pick the wrong config to display. Instead, we handle the
on_pre_move event and defer the update until after that event finishes
triggering.

The format of the display in the status bar is user configurable via a
new setting. A settings listener ensures that if the setting changes,
all of the windows immediately update.