Include config name in applied environment vars
In addition to including a variable that specifically tells a task that
Envault has applied an environment, also include a variable that
defines the name of the configuration file that was applied.

This has been useful for debugging purposes when done in an ad-hoc way,
so it makes sense to include it "for real".