Fix a bug in logging and enhance the var display
This commit fixes a bug in the logging of fetch failures, which try to
log the result from the request that failed and cause a failure
themselves because the logger would erroneously think that it should be
trying to do a format (since the error return is JSON).

In addition, augment the command for showing variables so that selecting
the item with the primary modifier key for your platform held causes
the value of the variable to be copied to the clipboard.